+++ extra/yolov3_test.py
@@ -15,16 +15,13 @@ def test_conv2d():
 
   # Torch output
   tconv = torch.nn.Conv2d(16, 33, 3, stride=2)
-  def fromNumpy(ndarray): # have to use this since PyTorch on M1 doesn't have NumPy support… smh
+  def fromNumpy(ndarray): # have to use this since PyTorch on M1 doesn't have NumPy support yet… smh
     return torch.Tensor(ndarray.tolist())
 
   tconv.bias = torch.nn.Parameter(fromNumpy(conv.biases.cpu().data).squeeze())
   tconv.weight = torch.nn.Parameter(fromNumpy(conv.weights.cpu().data))
   tout = tconv(fromNumpy(x.cpu().data))
-  #print("Output (tinygrad)")
-  #print(out.cpu().data[0][0][0])
-  #print("Output (Torch)")
-  #print(tout[0][0][0])
+
   np.testing.assert_allclose(out.cpu().data, tout.detach().numpy(), rtol=5e-2)
 
 def test_upsample():
@@ -32,8 +29,6 @@ def test_upsample():
   tx = torch.arange(1,5).view(1,1,2,2).float()
   m = Upsample(scale_factor=2, mode='nearest')
   tm = torch.nn.Upsample(scale_factor=2, mode='nearest')
-  #print("Output (tinygrad)")
-  #print(m(x).cpu().data)
   np.testing.assert_allclose(m(x).cpu().data, tm(tx).detach().numpy(), rtol=5e-5)
 
 np.random.seed(1337)
